Moved a few things around and abstracted moderative operations to their own methods.
