+++ src/Management/PunishmentManager.cs
@@ -1,10 +1,9 @@
 ﻿using sodoffmmo.Core;
-using sodoffmmo.Data;
 using System.Text.Json;
 
 namespace sodoffmmo.Management;
 
-public class PunishmentManager {
+public static class PunishmentManager {
     // VikingId, Reason (nullable)
     public static readonly Dictionary<int, string?> MutedList;
     public static readonly Dictionary<int, string?> BannedList;
@@ -35,69 +34,97 @@ public class PunishmentManager {
              WriteIndented = true
         }));
     }
+
+    /// <summary>
+    /// Mute a user.
+    /// </summary>
+    /// <param name="target">User to mute - can be null, but id should be set in that case.</param>
+    /// <param name="id">Optional vikingid of the user to mute.</param>
+    /// <returns>False if already muted, otherwise true.</returns>
+    public static bool Mute(this Client? target, int? id = null) {
+        if (target != null) {
+            id ??= target.PlayerData.VikingId;
+            if (target.Muted && id == null) return false;
+            target.Muted = true;
+        }
+
+        if (id != null) {
+            if (MutedList.ContainsKey((int)id)) return false;
+            MutedList.Add((int)id, target?.PlayerData.DiplayName);
+            SaveMutedBanned();
+        }
+
+        return true;
+    }
+
+    /// <summary>
+    /// Unmute a user.
+    /// </summary>
+    /// <param name="target">User to unmute - can be null, but id should be set in that case.</param>
+    /// <param name="id">Optional vikingid of the user to unmute.</param>
+    /// <returns>False if the user isn't muted, otherwise true.</returns>
+    public static bool Unmute(this Client? target, int? id=null) {
+        if (target != null) {
+            if (!target.Muted) return false;
+            target.Muted = false;
+            id ??= target.PlayerData.VikingId;
+        }
+
+        if (id != null) {
+            if (!MutedList.Remove((int)id)) return false;
+            SaveMutedBanned();
+        }
+
+        return true;
+    }
+
+    /// <summary>
+    /// Ban a user.
+    /// </summary>
+    /// <param name="target">User to ban - can be null, but id should be set in that case.</param>
+    /// <param name="id">Optional vikingid of the user to ban.</param>
+    /// <returns>False if already banned, otherwise true.</returns>
+    public static bool Ban(this Client? target, int? id=null) {
+        if (target != null) {
+            if (target.Banned) return false;
+            target.Banned = true;
+            target.SetRoom(BanRoom);
+            id ??= target.PlayerData.VikingId;
+        }
+
+        if (id != null) {
+            if (BannedList.ContainsKey((int)id)) return false;
+            BannedList.Add((int)id, target?.PlayerData.DiplayName);
+            SaveMutedBanned();
+        }
+
+        return true;
+    }
+
+    /// <summary>
+    /// Unban a user.
+    /// </summary>
+    /// <param name="target">User to unban - can be null, but id should be set in that case.</param>
+    /// <param name="id">Optional vikingid of the user to unban.</param>
+    /// <returns>False if the user isn't banned, otherwise true.</returns>
+    public static bool Unban(this Client? target, int? id=null) {
+        if (target != null) {
+            if (!target.Banned) return false;
+            target.Banned = false;
+            target.Disconnect(); // Client will handle relogging.
+            id ??= target.PlayerData.VikingId;
+        }
+
+        if (id != null) {
+            if (!BannedList.Remove((int)id)) return false;
+            SaveMutedBanned();
+        }
+
+        return true;
+    }
 }
 
 internal sealed class MutedBannedData {
     public Dictionary<int, string?>? MutedList { get; set; }
     public Dictionary<int, string?>? BannedList { get; set; }
-}
-
-public class BanRoom : Room {
-    public BanRoom(string name) : base(name, name) {}
-
-    public void UpdateClient(Client client) {
-        client.Send(RespondJoinRoom());
-    }
-
-    public override void Send(NetworkPacket packet, Client? skip = null) {
-        // Do nothing.
-    }
-
-    public override NetworkPacket RespondJoinRoom() {
-        NetworkObject obj = new();
-        NetworkArray roomInfo = new();
-        roomInfo.Add(Id);
-        roomInfo.Add(Name); // Room Name
-        roomInfo.Add(Group); // Group Name
-        roomInfo.Add(true); // is game
-        roomInfo.Add(false); // is hidden
-        roomInfo.Add(false); // is password protected
-        roomInfo.Add((short)0); // player count
-        roomInfo.Add((short)1); // max player count
-        roomInfo.Add(new NetworkArray()); // empty roomvars
-        roomInfo.Add((short)0); // spectator count
-        roomInfo.Add((short)0); // max spectator count
-
-        obj.Add("r", roomInfo);
-        obj.Add("ul", new NetworkArray()); // Empty userlist.
-
-        NetworkPacket packet = NetworkObject.WrapObject(0, 4, obj).Serialize();
-        packet.Compress();
-
-        return packet;
-    }
-
-    public override NetworkPacket SubscribeRoom() {
-        NetworkObject obj = new();
-        NetworkArray list = new();
-
-        NetworkArray r1 = new();
-        r1.Add(Id);
-        r1.Add(Name); // Room Name
-        r1.Add(Group); // Group Name
-        r1.Add(true);
-        r1.Add(false);
-        r1.Add(false);
-        r1.Add((short)0); // player count
-        r1.Add((short)1); // max player count
-        r1.Add(new NetworkArray());
-        r1.Add((short)0);
-        r1.Add((short)0);
-
-        list.Add(r1);
-
-        obj.Add("rl", list);
-        obj.Add("g", Group);
-        return NetworkObject.WrapObject(0, 15, obj).Serialize();
-    }
 }
